Session 3/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai>=1.3.0
pillow==10.1.0
